Refactor broker (#164)
