+++ broker/config/server.js
@@ -7,7 +7,8 @@ dotenv.config();
 // NOTE: login/password settings should not be defaulted.
 //
 module.exports = {
-    port: process.env.PORT || 3000,
-    fhirUrlBase: process.env.FHIR_URL_BASE || "http://hapi:8080/hapi-fhir-jpaserver/fhir",
-    mirthUrl: process.env.MIRTH_URL || "http://localhost:8444/"
-}
+  port: process.env.PORT || 3000,
+  fhirUrlBase:
+    process.env.FHIR_URL_BASE || "http://hapi:8080/hapi-fhir-jpaserver/fhir",
+  mirthUrl: process.env.MIRTH_URL || "http://localhost:8444/",
+};
