+++ broker/endpoints/Location.js
@@ -1,0 +1,31 @@
+const axios = require("../services/axiosInstance.js");
+const Location = require("../models/Location");
+
+exports.read = (req, res) => {
+  axios
+    .get(`${req.url}`)
+    .then((response) => {
+      res.json(response.data);
+    })
+    .catch((e) => {
+      res.status(400).json({
+        error: e.response ? e.response.data : e.message,
+      });
+    });
+};
+
+exports.create = (req, res) => {
+  let loc = req.body.Location;
+  let resource = Location.toFHIR(loc);
+
+  axios
+    .post(`/Location`, resource)
+    .then((response) => {
+      res.json(response.data);
+    })
+    .catch((e) => {
+      res.status(400).json({
+        error: e.response ? e.response.data : e.message,
+      });
+    });
+};
