+++ broker/models/Appointment.js
@@ -17,27 +17,6 @@ Appointment {
   ]
   slot : string "Slot/id" // note that this is not needed until scheduling is implemented
 }
-{
-  "Appointment":
-    {
-      "status":"booked",
-      "participant":[
-        {
-          "type":"patient",
-          "actor":"Patient/example"
-        },
-        {
-          "type":"pratitioner",
-          "actor":"Practitioner/f201"
-        },
-        {
-          "type":"patient",
-          "actor":"Patient/102"
-        }
-      ],
-      "slot":"Slot/1253"
-    }
-}
 */
 exports.toFHIR = function (appt) {
   let resource = {
