+++ broker/models/Encounter.js
@@ -8,18 +8,6 @@ Encounter {
   end : string "YYYY, YYYY-MM, YYYY-MM-DD or YYYY-MM-DDThh:mm:ss+zz:zz"
   location : string "Location/id"
 }
-{
-  "Encounter": {
-    "status" : "planned",
-    "class" : "FLD",
-    "subject" : "Patient/1",
-    "appointment" : "Appointment/153",
-    "start" : "2020-09-17T05:10:19+00:00",
-    "end" : "2020-09-18T05:15:19+00:00",
-    "location" : "Location/304",
-    "serviceProvider" : "Organization/252"
-  }
-}
 */
 const CLASS_SYSTEM = "http://terminology.hl7.org/CodeSystem/v3-ActCode";
 exports.toFHIR = function (encounter) {
@@ -34,9 +22,11 @@ exports.toFHIR = function (encounter) {
     subject: {
       reference: encounter.subject,
     },
-    appointment: {
-      reference: encounter.appointment,
-    },
+    appointment: [
+      {
+        reference: encounter.appointment,
+      },
+    ],
     period: {
       start: encounter.start,
       end: encounter.end,
@@ -54,4 +44,4 @@ exports.toFHIR = function (encounter) {
   };
 
   return resource;
-}
+};
