+++ broker/models/Practitioner.js
@@ -17,26 +17,6 @@ Practitioner {
   qualificationCode : enum (BN, CANP, CER, CMA, CNP, CNS, CRN, EMT, RN, RMA, PN, RPH)
   }
 }
-{
-  "Practitioner":
-    {
-      "family" : "George",
-      "given" : "James",
-      "suffix" : "Sr.",
-      "phone" : "(010)000-000",
-      "email" : "email@site.com",
-      "gender": "male",
-      "birthDate": "2000-01-01",
-      "address": {
-        "line": "97 Walnut St",
-        "city": "New York City",
-        "state": "NY",
-        "postalCode": "14623",
-        "country": "USA"
-      },
-      "qualificationCode": "CER"
-   }
-}
 */
 
 exports.toFHIR = function (prt) {
